+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - #3091, Broken link in OpenAPI description `externalDocs` - @salim-b
  - #3659, Embed One-to-One relationship with different column order properly - @wolfgangwalther
  - #3504, Remove `format` from `rowFilter` parameters in OpenAPI - @dantheman2865
+ - #3640, Prevent spread embedding to allow aggregates when they are disabled - @laurenceisla
 
 ### Changed